improve object typing
